+++ src/index.js
@@ -1,3 +1,3 @@
-export {load, supportedExtensions} from './load.js'
+export {load, loadAsync, supportedExtensions} from './load.js'
 export {parse} from './parse.js'
 export {stringify} from './stringify.js'