+++ changelog.md
@@ -4,7 +4,7 @@
 
  * Remove filename option from parse
  * Works in the browser
-
+ * Rename dist/index.js dist/index.cjs
 
 ## 2.1.0